stat manager:

+++ src/com/kai/game/entities/DoesCombat.java
@@ -1,7 +1,7 @@
 package com.kai.game.entities;
 
 public interface DoesCombat {
-    void takeDamage(int amount);
+    void takeDamage(double amount);
     void heal(double amount);
     void attack(Entity target);
     void attack(Entity target, int ovrDamage);
